404 for all route with an id

+++ src/Controller/Back/UserController.php
@@ -61,8 +61,14 @@ class UserController extends AbstractController
     /**
      * @Route("/{id}", name="app_user_show", methods={"GET"})
      */
-    public function show(User $user): Response
+    public function show(User $user = null): Response
     {
+        //404?
+        if($user === null)
+        {
+            return $this->json(['error' => 'Utilisateur non trouvé.'], Response::HTTP_NOT_FOUND);
+        }
+
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -71,8 +77,14 @@ class UserController extends AbstractController
     /**
      * @Route("/{id}/edit", name="app_user_edit", methods={"GET", "POST"})
      */
-    public function edit(Request $request, User $user, UserRepository $userRepository): Response
+    public function edit(Request $request, User $user = null, UserRepository $userRepository): Response
     {
+        //404?
+        if($user === null)
+        {
+            return $this->json(['error' => 'Utilisateur non trouvé.'], Response::HTTP_NOT_FOUND);
+        }
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -91,8 +103,14 @@ class UserController extends AbstractController
     /**
      * @Route("/{id}", name="app_user_delete", methods={"POST"})
      */
-    public function delete(Request $request, User $user, UserRepository $userRepository): Response
+    public function delete(Request $request, User $user = null, UserRepository $userRepository): Response
     {
+        //404?
+        if($user === null)
+        {
+            return $this->json(['error' => 'Utilisateur non trouvé.'], Response::HTTP_NOT_FOUND);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user, true);
         }
